optimise `push_pair!`

+++ src/setup.jl
@@ -735,7 +735,7 @@ function System(coord_file::AbstractString,
                 error("atom $ai type $atype has charge missing")
             end
         end
-        push!(atoms_abst, Atom(index=ai, atom_type=ati, mass=at.mass, charge=ch,
+        push!(atoms_abst, Atom(index=Int32(ai), atom_type=Int32(ati), mass=at.mass, charge=ch,
                                σ=at.σ, ϵ=at.ϵ, λ=T(1.0)))
 
         if !ismissing(at.σ14)
